New feature, finally got around to adding the first time locales reconfig option prior to first du.
this runs only for sidux users, since debian doesn't turn on so many locales by default.

this is a one time option, and required updating the locales function to handle yes/no options
to allow for smooth flow.

I've been meaning to get to this one for a long time.