Update to inxi installer in package install,. forgot to have it check for apt package version.
Now it will install the apt package if available, the script version if not. Also forces an update
of inxi no matter what after, to install man pages for script installs, and just to make sure the distro
version is new. Siduction disables U last I checked, so this should only matter to distros that don't change that default.

Fixed google chrome installer to create google-chrome-unstable.list file for debian sid/testing, that's the file
that chrome insists on creating whether or not there's a source listed for it already in apt sources.